examples/WaterMarkingWebsite/SERVER/requirements.txt
Django==3.1.4
django-cors-headers==3.6.0
pymongo==3.11.2
requests==2.32.2
passlib==1.7.4
librosa==0.10.2.post1
pyjwt
